undo moving folder (breakin)
