+++ src/pages/compare-state/components/Tree/components/NormalTree/index.jsx
@@ -1,6 +1,6 @@
 import PropTypes from "prop-types";
 
-import { getIndent } from "pages/history/components/comparison/utils";
+import { getIndent } from "pages/compare-state/utils";
 
 export const NormalTree = (props) => {
   let {
