+++ src/pages/compare-state/index.ts
@@ -1,0 +1,1 @@
+export {CompareState} from "./compareState"
