+++ src/pages/compare-state/styles.js
