+++ src/pages/compare-state/utils/getType.js
