+++ src/pages/compare-state/utils/index.js
