+++ src/pages/compare-state/utils/isArray.js
