+++ src/pages/compare-state/utils/isObject.js
