+++ src/pages/compare-state/utils/needFormat.js
