+++ src/pages/history/history.tsx
@@ -16,7 +16,6 @@ import { ContentHeader } from "shared/components/content-header";
 import { useDispatch } from "react-redux";
 
 import * as S from "./styles";
-import { Comparison } from "./components/comparison/comparison";
 
 export const History: React.FC<{}> = () => {
   const params = useParams();
@@ -69,7 +68,6 @@ export const History: React.FC<{}> = () => {
         <S.Table columnData={table.columnData} rows={table.rows} isCollapse />
       </S.TableContainer>
 
-      <Comparison></Comparison>
     </S.Wrapper>
   );
 };
